pyproject.toml
[tool.poetry]
name = "llava-fasthtml"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.3.3"
boto3 = "^1.34.158"
python-dotenv = "^1.0.1"
pytailwindcss = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
